plugins/mkdocs-llm-autodoc/requirements.txt
# MkDocs LLM AutoDoc Plugin Requirements

# Core dependencies
mkdocs>=1.4.0
pyyaml>=6.0

# LLM Providers
anthropic>=0.18.0
openai>=1.0.0

# Code Parsing
tree-sitter>=0.21.0
tree-sitter-cpp>=0.21.0

# Utilities
jinja2>=3.0.0
watchdog>=3.0.0
tqdm>=4.65.0

# Optional: For development
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
